+++ NEWS.md
@@ -4,6 +4,8 @@
 
 ### New
 * length_sort(): Sorts vector elements by character length.
+* length_order(): Sorts data.frames by character length of elements in specified
+column(s).
 
 
 # DO.utils 0.2.7